Wed, Aug 23, 2023, 5:28 PM +02:00

+++ credit/credit.c
@@ -107,9 +107,9 @@ int main(void)
                 }
 
                 long lastNumsCopy = last2Nums;
-                int tempReplaced1 = lastNumsCopy % 10;
+                long tempReplaced1 = lastNumsCopy % 10;
                 lastNumsCopy = lastNumsCopy / 10;
-                int tempReplaced2 = lastNumsCopy % 10;
+                long tempReplaced2 = lastNumsCopy % 10;
